Fix handleEditor
added two variable to save the frame when freeze effect is applied and a boolean variable to save the flip effect status in that moment. These are used to tell the editor controller whether to flip or not the imageview, keeping in mind that if i pressed the flip button while the cam is frozen that shouldn't affect the effect of the preview imageview.