Removed toggleable
I think that the current class effect is a better generalization of itself as concept. Created an overload for apply() so that it replace toggle. apply(ImageView imageAffected) will be implemented by all Effect implementation according to its needs. LiveEffect is an empty class because it is only a conceptual division to distinguish normal effects from live effects. apply() method still exists because it must to exists into an interface called Applicable